+++ src/app/page.tsx
@@ -1,8 +1,8 @@
 import Header from "@/components/Header";
 import HeroSection from "@/components/Hero";
 import Usage from "@/components/Usage";
-import Banner from "@/components/Banner";
-import Banner2 from "@/components/Banner2";
+import JoinForm from "@/components/JoinForm";
+import SponsorForm from "@/components/SponsorForm";
 import { IconCloudDemo } from "@/components/IconCloudDemo";
 import Footer from "@/components/Footer";
 import { FileTreeDemo } from "@/components/FileTree";
@@ -15,11 +15,11 @@ export default function Home() {
 
       <Header />
       <HeroSection />
-      <Banner />
+      <JoinForm />
       <FileTreeDemo />
 
       <IconCloudDemo />
-      <Banner2 />
+      <SponsorForm />
       <Usage />
       <Footer />
     </>
